Query the OTN Geoserver

Query the OTN Geoserver

Usage

otn_query(projects)

Arguments

projects

Character vector of OTN project codes for which you'd like project metadata. Prepended networks can be provided, but are not necessary.

Value

list of the "otn_resources_metadata_points" and "project_metadata" for the given projects

Examples

otn_query(c("EST", "FACT.SCDNRDFP", "ACT.MDBSB", "MDBSB"))
#> $otn_resources_metadata_points
#>                                                             FID collectioncode
#>                                                          <char>         <char>
#> 1: otn_resources_metadata_points.fid-66c0a928_190832eaa66_-5911            EST
#> 2: otn_resources_metadata_points.fid-66c0a928_190832eaa66_-572c       SCDNRDFP
#> 3: otn_resources_metadata_points.fid-66c0a928_190832eaa66_-5655          MDBSB
#>    report
#>    <lgcl>
#> 1:     NA
#> 2:     NA
#> 3:     NA
#>                                                                                                               resource_full_name
#>                                                                                                                           <char>
#> 1:                                                                        Shedd Aquarium Bahamas Sharks and Ray Research Program
#> 2:                                                                                    Diadromous Fishes statewide movement in SC
#> 3: Influence of Turbine Construction Noise on Black Sea Bass Displacement and Physiological Condition in the MD Wind Energy Area
#>          ocean seriescode    status collaborationtype totalrecords    id
#>         <char>     <char>    <char>            <char>       <lgcl> <int>
#> 1: NW ATLANTIC  OTNGlobal   ongoing              Data           NA   312
#> 2: NW ATLANTIC       FACT   ongoing        Deployment           NA   836
#> 3: NE ATLANTIC        ACT completed              Data           NA  1050
#>                                                                          the_geom
#>                                                                            <char>
#> 1: POLYGON ((-76.85 23.62, -76.85 24.83, -75.9 24.83, -75.9 23.62, -76.85 23.62))
#> 2:     POLYGON ((-82.46 25.41, -82.46 35.99, -78 35.99, -78 25.41, -82.46 25.41))
#> 3: POLYGON ((-79.53 33.88, -79.53 42.26, -68.1 42.26, -68.1 33.88, -79.53 33.88))
#> 
#> $project_metadata
#>                                                FID collectioncode    id
#>                                             <char>         <char> <int>
#> 1: project_metadata.fid-66c0a928_190832eaa66_-5541           TEST    37
#> 2: project_metadata.fid-66c0a928_190832eaa66_-5450           TEST   279
#> 3: project_metadata.fid-66c0a928_190832eaa66_-5426           TEST   322
#>    seriescode collaborationtype shortname    longname citation abstract
#>        <char>            <lgcl>    <char>      <char>   <lgcl>   <lgcl>
#> 1:     GLOBAL                NA      TEST Test Series       NA       NA
#> 2:     GLOBAL                NA      TEST Test Series       NA       NA
#> 3:     GLOBAL                NA      TEST Test Series       NA       NA
#>    institutioncode  ocean country  state local_area locality westbl eastbl
#>             <lgcl> <char>  <lgcl> <lgcl>     <lgcl>   <lgcl>  <int>  <int>
#> 1:              NA GLOBAL      NA     NA         NA       NA   -180    180
#> 2:              NA GLOBAL      NA     NA         NA       NA   -180    180
#> 3:              NA GLOBAL      NA     NA         NA       NA   -180    180
#>    southbl northbl  status  usage website      sdate   edate       node
#>      <int>   <int>  <char> <lgcl>  <lgcl>     <IDat>  <char>     <char>
#> 1:     -90      90 ongoing     NA      NA 2000-01-01 ongoing OTN-Global
#> 2:     -90      90 ongoing     NA      NA 2000-01-01 ongoing        nep
#> 3:     -90      90 ongoing     NA      NA       <NA>         OTN-Global
#>    database       db_location datacentre
#>      <char>            <char>     <lgcl>
#> 1:      otn db.oceantrack.org         NA
#> 2:  nepnode db.oceantrack.org         NA
#> 3:      otn db.oceantrack.org         NA
#>                                                    the_geom
#>                                                      <char>
#> 1: POLYGON ((-180 -90, -180 90, 180 90, 180 -90, -180 -90))
#> 2: POLYGON ((-180 -90, -180 90, 180 90, 180 -90, -180 -90))
#> 3: POLYGON ((-180 -90, -180 90, 180 90, 180 -90, -180 -90))
#>